refactor(desktop): extract link preview helpers
